remove snowpark_package
